sync

+++ app/code/Magento/Email/Controller/Adminhtml/Email/Template.php
@@ -18,8 +18,6 @@
  * versions in the future. If you wish to customize Magento for your
  * needs please refer to http://www.magentocommerce.com for more information.
  *
- * @category    Magento
- * @package     Magento_Email
  * @copyright   Copyright (c) 2014 X.commerce, Inc. (http://www.magentocommerce.com)
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
@@ -28,8 +26,6 @@ namespace Magento\Email\Controller\Adminhtml\Email;
 /**
  * System Template admin controller
  *
- * @category   Magento
- * @package    Magento_Email
  * @author      Magento Core Team <core@magentocommerce.com>
  */
 class Template extends \Magento\Backend\App\Action
@@ -37,18 +33,16 @@ class Template extends \Magento\Backend\App\Action
     /**
      * Core registry
      *
-     * @var \Magento\Registry
+     * @var \Magento\Framework\Registry
      */
     protected $_coreRegistry = null;
 
     /**
      * @param \Magento\Backend\App\Action\Context $context
-     * @param \Magento\Registry $coreRegistry
+     * @param \Magento\Framework\Registry $coreRegistry
      */
-    public function __construct(
-        \Magento\Backend\App\Action\Context $context,
-        \Magento\Registry $coreRegistry
-    ) {
+    public function __construct(\Magento\Backend\App\Action\Context $context, \Magento\Framework\Registry $coreRegistry)
+    {
         $this->_coreRegistry = $coreRegistry;
         parent::__construct($context);
     }
@@ -114,9 +108,13 @@ class Template extends \Magento\Backend\App\Action
 
         $this->_title->add($template->getId() ? $template->getTemplateCode() : __('New Template'));
 
-        $this->_addContent($this->_view->getLayout()
-            ->createBlock('Magento\Email\Block\Adminhtml\Template\Edit', 'template_edit')
-            ->setEditMode((bool)$this->getRequest()->getParam('id'))
+        $this->_addContent(
+            $this->_view->getLayout()->createBlock(
+                'Magento\Email\Block\Adminhtml\Template\Edit',
+                'template_edit'
+            )->setEditMode(
+                (bool)$this->getRequest()->getParam('id')
+            )
         );
         $this->_view->renderLayout();
     }
@@ -139,20 +137,28 @@ class Template extends \Magento\Backend\App\Action
         }
 
         try {
-            $template->setTemplateSubject($request->getParam('template_subject'))
-                ->setTemplateCode($request->getParam('template_code'))
-                ->setTemplateText($request->getParam('template_text'))
-                ->setTemplateStyles($request->getParam('template_styles'))
-                ->setModifiedAt($this->_objectManager->get('Magento\Stdlib\DateTime\DateTime')->gmtDate())
-                ->setOrigTemplateCode($request->getParam('orig_template_code'))
-                ->setOrigTemplateVariables($request->getParam('orig_template_variables'));
+            $template->setTemplateSubject(
+                $request->getParam('template_subject')
+            )->setTemplateCode(
+                $request->getParam('template_code')
+            )->setTemplateText(
+                $request->getParam('template_text')
+            )->setTemplateStyles(
+                $request->getParam('template_styles')
+            )->setModifiedAt(
+                $this->_objectManager->get('Magento\Framework\Stdlib\DateTime\DateTime')->gmtDate()
+            )->setOrigTemplateCode(
+                $request->getParam('orig_template_code')
+            )->setOrigTemplateVariables(
+                $request->getParam('orig_template_variables')
+            );
 
             if (!$template->getId()) {
                 $template->setTemplateType(\Magento\Email\Model\Template::TYPE_HTML);
             }
 
             if ($request->getParam('_change_type_flag')) {
-                $template->setTemplateType(\Magento\App\TemplateTypesInterface::TYPE_TEXT);
+                $template->setTemplateType(\Magento\Framework\App\TemplateTypesInterface::TYPE_TEXT);
                 $template->setTemplateStyles('');
             }
 
@@ -161,12 +167,15 @@ class Template extends \Magento\Backend\App\Action
             $this->messageManager->addSuccess(__('The email template has been saved.'));
             $this->_redirect('adminhtml/*');
         } catch (\Exception $e) {
-            $this->_objectManager->get('Magento\Backend\Model\Session')
-                ->setData('email_template_form_data', $request->getParams());
+            $this->_objectManager->get(
+                'Magento\Backend\Model\Session'
+            )->setData(
+                'email_template_form_data',
+                $request->getParams()
+            );
             $this->messageManager->addError($e->getMessage());
             $this->_forward('new');
         }
-
     }
 
     /**
@@ -179,22 +188,34 @@ class Template extends \Magento\Backend\App\Action
         $template = $this->_initTemplate('id');
         if ($template->getId()) {
             try {
-                $template->delete();
-                 // display success message
-                $this->messageManager->addSuccess(__('The email template has been deleted.'));
-                // go to grid
-                $this->_redirect('adminhtml/*/');
+                // check if the template is currently used
+                if (count($template->getSystemConfigPathsWhereUsedCurrently()) == 0) {
+                    $template->delete();
+                    // display success message
+                    $this->messageManager->addSuccess(__('The email template has been deleted.'));
+                    $this->_objectManager->get('Magento\Framework\App\ReinitableConfig')->reinit();
+                    // go to grid
+                    $this->_redirect('adminhtml/*/');
+                    return;
+                }
+                // display error  message
+                $this->messageManager->addError(__('The email template is currently being used.'));
+                // redirect to edit form
+                $this->_redirect('adminhtml/*/edit', array('id' => $template->getId()));
                 return;
-            } catch (\Magento\Core\Exception $e) {
+            } catch (\Magento\Framework\Model\Exception $e) {
                 $this->messageManager->addError($e->getMessage());
             } catch (\Exception $e) {
                 $this->messageManager->addError(
                     __('An error occurred while deleting email template data. Please review log and try again.')
                 );
-                $this->_objectManager->get('Magento\Logger')->logException($e);
+                $this->_objectManager->get('Magento\Framework\Logger')->logException($e);
                 // save data in session
-                $this->_objectManager->get('Magento\Backend\Model\Session')
-                    ->setFormData($this->getRequest()->getParams());
+                $this->_objectManager->get(
+                    'Magento\Backend\Model\Session'
+                )->setFormData(
+                    $this->getRequest()->getParams()
+                );
                 // redirect to edit form
                 $this->_redirect('adminhtml/*/edit', array('id' => $template->getId()));
                 return;
@@ -213,8 +234,13 @@ class Template extends \Magento\Backend\App\Action
      */
     public function previewAction()
     {
-        $this->_view->loadLayout('systemPreview');
-        $this->_view->renderLayout();
+        try {
+            $this->_view->loadLayout('systemPreview');
+            $this->_view->renderLayout();
+        } catch (\Exception $e) {
+            $this->messageManager->addError(__('An error occurred. The email template can not be opened for preview.'));
+            $this->_redirect('adminhtml/*/');
+        }
     }
 
     /**
@@ -224,6 +250,7 @@ class Template extends \Magento\Backend\App\Action
      */
     public function defaultTemplateAction()
     {
+        $this->_view->loadLayout();
         $template = $this->_initTemplate('id');
         $templateCode = $this->getRequest()->getParam('code');
         try {
@@ -238,7 +265,7 @@ class Template extends \Magento\Backend\App\Action
                 $this->_objectManager->get('Magento\Core\Helper\Data')->jsonEncode($template->getData())
             );
         } catch (\Exception $e) {
-            $this->_objectManager->get('Magento\Logger')->logException($e);
+            $this->_objectManager->get('Magento\Framework\Logger')->logException($e);
         }
     }
